+++ src/loader.rs
@@ -396,7 +396,8 @@ fn load_map(
 
     // Order of the differents layers in the .TMX file is important:
     // a layer appearing last in the .TMX should appear "on top" of previous layers
-    let mut offset_z = 0.;
+    // Start with a negative offset so the upper layer will be at Z-offset = 0
+    let mut offset_z = tiled_map.map.layers().len() as f32 * (-100.0);
 
     // Once materials have been created/added we need to then create the layers.
     for (layer_index, layer) in tiled_map.map.layers().enumerate() {